requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.0
kiwisolver==1.4.7
matplotlib==3.9.2
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
readerwriterlock==1.0.9
six==1.16.0
texttable==1.7.0
typing_extensions==4.12.2
tzdata==2024.2
